--- src/main/java/com/nbcampif/ifstagram/global/jwt/RefreshTokenJpaRepository.java
@@ -1,8 +1,0 @@
-package com.nbcampif.ifstagram.global.jwt;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface RefreshTokenJpaRepository extends JpaRepository<RefreshTokenEntity, Long> {
-
-}
