+++ src/main/java/com/nbcampif/ifstagram/global/jwt/repository/RefreshTokenJpaRepository.java
@@ -1,0 +1,16 @@
+package com.nbcampif.ifstagram.global.jwt.repository;
+
+import com.nbcampif.ifstagram.global.jwt.RefreshTokenEntity;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+public interface RefreshTokenJpaRepository extends JpaRepository<RefreshTokenEntity, Long> {
+
+  Optional<RefreshTokenEntity> findByToken(String refreshToken);
+
+  void deleteByUserId(Long userId);
+
+  void deleteByToken(String refreshToken);
+
+}
